# public/swagger.json
{
	"components": {
...
		"schemas": {
			"Types.ObjectId": {
				"type": "string"
			},
			"IUsers": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"email": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"profilePicture": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"gender": {
						"type": "string"
					},
					"DOB": {
						"type": "string",
						"format": "date-time"
					},
					"address": {
						"items": {
							"properties": {
								"manipolicity": {
									"type": "string"
								},
								"government": {
									"type": "string"
								},
								"country": {
									"type": "string"
								}
							},
							"required": [
								"manipolicity",
								"government",
								"country"
							],
							"type": "object"
						},
						"type": "array"
					},
					"verified": {
						"type": "string",
						"enum": [
							"notSent",
							"pending",
							"verified"
						]
					},
					"status": {
						"type": "string",
						"enum": [
							"inActive",
							"active",
							"suspended",
							"lost",
							"deleted"
						]
					},
					"accountType": {
						"type": "string",
						"enum": [
							"PT",
							"EM",
							"PA"
						]
					},
					"lastLoginDate": {
						"type": "string",
						"format": "date-time"
					},
					"accountSetting": {
						"additionalProperties": false,
						"type": "object"
					},
					"languages": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"maritalStatus": {
						"type": "string",
						"enum": [
							"married",
							"single",
							"divorced",
							"widow"
						]
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"phoneNumber",
					"password",
					"profilePicture",
					"firstName",
					"lastName",
					"gender",
					"DOB",
					"verified",
					"status",
					"accountType",
					"lastLoginDate",
					"accountSetting",
					"languages",
					"maritalStatus"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IClients": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						]
					},
					"uID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUsers"
							}
						]
					},
					"preferredServiceType": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"type": "string",
									"enum": [
										"online",
										"home",
										"office"
									]
								},
								"type": "array"
							}
						]
					},
					"diseases": {
						"type": "string"
					},
					"preferences": {
						"additionalProperties": false,
						"type": "object"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"preferredServiceType",
					"diseases",
					"preferences"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISessions": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"sessionType": {
						"type": "string",
						"enum": [
							"group",
							"individual"
						]
					},
					"serviceProvidersID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IServiceProviders"
							}
						]
					},
					"clientsIDs": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IClients"
								},
								"type": "array"
							}
						]
					},
					"name": {
						"type": "string"
					},
					"details": {
						"type": "string"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"duration": {
						"type": "number",
						"format": "double"
					},
					"serviceType": {
						"type": "string",
						"enum": [
							"Online",
							"Home",
							"Office"
						]
					},
					"location": {
						"additionalProperties": false,
						"type": "object"
					},
					"attachments": {
						"properties": {
							"attachmentType": {
								"type": "string"
							},
							"attachmentName": {
								"type": "string"
							},
							"attachmentUrl": {
								"type": "string"
							}
						},
						"required": [
							"attachmentType",
							"attachmentName",
							"attachmentUrl"
						],
						"type": "object"
					},
					"requirements": {
						"type": "string"
					},
					"ratings": {
						"items": {
							"properties": {
								"ratingDate": {
									"type": "string",
									"format": "date-time"
								},
								"ratingValue": {
									"type": "string"
								},
								"raterUID": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"$ref": "#/components/schemas/Types.ObjectId"
										},
										{
											"$ref": "#/components/schemas/IUsers"
										}
									]
								}
							},
							"required": [
								"ratingDate",
								"ratingValue",
								"raterUID"
							],
							"type": "object"
						},
						"type": "array"
					},
					"reviews": {
						"items": {
							"properties": {
								"reviewDate": {
									"type": "string",
									"format": "date-time"
								},
								"reviewDetails": {
									"type": "string"
								},
								"reviewerUID": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"$ref": "#/components/schemas/Types.ObjectId"
										},
										{
											"$ref": "#/components/schemas/IUsers"
										}
									]
								}
							},
							"required": [
								"reviewDate",
								"reviewDetails",
								"reviewerUID"
							],
							"type": "object"
						},
						"type": "array"
					},
					"sessionFee": {
						"type": "number",
						"format": "double"
					},
					"payments": {
						"properties": {
							"amount": {
								"type": "number",
								"format": "double"
							},
							"payerID": {
								"anyOf": [
									{
										"type": "string"
									},
									{
										"$ref": "#/components/schemas/Types.ObjectId"
									},
									{
										"$ref": "#/components/schemas/IUsers"
									}
								]
							},
							"paymentMethod": {
								"type": "string"
							},
							"discount": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"amount",
							"payerID",
							"paymentMethod",
							"discount"
						],
						"type": "object"
					},
					"status": {
						"type": "string",
						"enum": [
							"initiated",
							"agreed",
							"canceled",
							"finished"
						]
					},
					"doctorReferral": {
						"type": "string"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"sessionType",
					"name",
					"details",
					"startDate",
					"duration",
					"serviceType",
					"location",
					"requirements",
					"sessionFee",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IAgreements": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						]
					},
					"name": {
						"type": "string"
					},
					"parties": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"details": {
						"type": "string"
					},
					"attachments": {
						"items": {
							"additionalProperties": false,
							"type": "object"
						},
						"type": "array"
					},
					"reminder": {
						"type": "boolean"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"parties",
					"startDate",
					"endDate",
					"details",
					"reminder"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Partial_IClients_": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						]
					},
					"uID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUsers"
							}
						]
					},
					"preferredServiceType": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"type": "string",
									"enum": [
										"online",
										"home",
										"office"
									]
								},
								"type": "array"
							}
						]
					},
					"diseases": {
						"type": "string"
					},
					"preferences": {
						"additionalProperties": false,
						"type": "object"
					},
					"url": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"FlattenMaps_T_": {
				"properties": {},
				"type": "object"
			},
			"Document_any.any.any_": {
				"$ref": "#/components/schemas/FlattenMaps_T_",
				"description": "Generic types for Document:\n*  T - the type of _id\n*  TQueryHelpers - Object with any helpers that should be mixed into the Query type\n*  DocType - the type of the actual Document created"
			},
			"Required_any_": {
				"properties": {},
				"type": "object",
				"description": "Make all properties in T required"
			},
...
			"IAlarms": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						]
					},
					"name": {
						"type": "string"
					},
					"referenceType": {
						"type": "string",
						"enum": [
							"Agreement",
							"Session",
							"Dispute",
							"ServiceProvider"
						]
					},
					"referenceID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/IAgreements"
							},
							{
								"$ref": "#/components/schemas/ISessions"
							},
							{
								"$ref": "#/components/schemas/IDisputes"
							},
							{
								"$ref": "#/components/schemas/IServiceProviders"
							}
						]
					},
					"frequencyUnit": {
						"type": "string",
						"enum": [
							"Days",
							"Hours"
						]
					},
					"frequency": {
						"type": "number",
						"format": "double"
					},
					"active": {
						"type": "boolean"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"referenceType",
					"frequencyUnit",
					"frequency",
					"active",
					"startDate",
					"endDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAnnouncements": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"referenceType": {
						"type": "string",
						"enum": [
							"Session",
							"Advertisement"
						]
					},
					"referenceID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/ISessions"
							},
							{
								"$ref": "#/components/schemas/INotifications"
							}
						]
					},
					"statues": {
						"type": "string",
						"enum": [
							"draft",
							"published"
						]
					},
					"topic": {
						"type": "string"
					},
					"details": {
						"type": "string"
					},
					"sentDate": {
						"type": "string",
						"format": "date-time"
					},
					"attachments": {
						"items": {
							"additionalProperties": false,
							"type": "object"
						},
						"type": "array"
					},
					"receiversUIDs": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IUsers"
								},
								"type": "array"
							}
						]
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"referenceType",
					"statues",
					"topic",
					"details",
					"sentDate",
					"attachments"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Partial_IAnnouncements_": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"referenceType": {
						"type": "string",
						"enum": [
							"Session",
							"Advertisement"
						]
					},
					"referenceID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/ISessions"
							},
							{
								"$ref": "#/components/schemas/INotifications"
							}
						]
					},
					"statues": {
						"type": "string",
						"enum": [
							"draft",
							"published"
						]
					},
					"topic": {
						"type": "string"
					},
					"details": {
						"type": "string"
					},
					"sentDate": {
						"type": "string",
						"format": "date-time"
					},
					"attachments": {
						"items": {
							"additionalProperties": false,
							"type": "object"
						},
						"type": "array"
					},
					"receiversUIDs": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IUsers"
								},
								"type": "array"
							}
						]
					},
					"url": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"Partial_IAlarms_": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						]
					},
					"name": {
						"type": "string"
					},
					"referenceType": {
						"type": "string",
						"enum": [
							"Agreement",
							"Session",
							"Dispute",
							"ServiceProvider"
						]
					},
					"referenceID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/IAgreements"
							},
							{
								"$ref": "#/components/schemas/ISessions"
							},
							{
								"$ref": "#/components/schemas/IDisputes"
							},
							{
								"$ref": "#/components/schemas/IServiceProviders"
							}
						]
					},
					"frequencyUnit": {
						"type": "string",
						"enum": [
							"Days",
							"Hours"
						]
					},
					"frequency": {
						"type": "number",
						"format": "double"
					},
					"active": {
						"type": "boolean"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"url": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/clients/{clientId}": {
			"get": {
				"operationId": "GetClient",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IClients"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "The requested client is not found"
					}
				},
				"description": "Get a client details",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "clientId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "6310750be8f4ab035351fb78"
					}
				]
			},
			"delete": {
				"operationId": "DeleteClient",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IClients"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "The requested client is not found"
					}
				},
				"description": "Delete a client",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "clientId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "6310750be8f4ab035351fb78"
					}
				]
			}
		},
		"/clients/create": {
			"post": {
				"operationId": "CreateClient",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IClients"
								},
								"examples": {
									"Example 1": {
										"value": {
											"uID": "63124dffdcf1e4974079a441",
											"preferredServiceType": [
												"online"
											],
											"diseases": "diseases",
											"preferences": {
												"pref": "pref test"
											}
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Create a client",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IClients"
							}
						}
					}
				}
			}
		},
		"/clients/update/{clientId}": {
			"put": {
				"operationId": "UpdateClient",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IClients"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Update a Client",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "clientId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "6310750be8f4ab035351fb78"
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_IClients_"
							},
							"example": "6310750be8f4ab035351fb78"
						}
					}
				}
			}
		},
...
		"/announcements/{announcementId}": {
			"get": {
				"operationId": "GetAnnouncement",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IAnnouncements"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "the requested Announcement in not found"
					}
				},
				"description": "Get a Announcement details",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "announcementId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteAnnouncement",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Document_any.any.any_"
										},
										{
											"$ref": "#/components/schemas/IAnnouncements"
										},
										{
											"$ref": "#/components/schemas/Required_any_"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "the requested announcement in not found"
					}
				},
				"description": "Delete a announcement",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "announcementId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "631330dbb5a0b3d3c3d6b0e7"
					}
				]
			}
		},
		"/announcements/create": {
			"post": {
				"operationId": "CreateAnnouncement",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAnnouncements"
								},
								"examples": {
									"Example 1": {
										"value": {
											"referenceType": "Session",
											"referenceID": "63131c361ea32f520cb28bd6",
											"statues": "draft",
											"topic": "new topic",
											"details": "The announcement contain info ",
											"sentDate": "2022-09-10T00:00:00.000Z",
											"attachments": [
												{
													"name": "personal information"
												}
											],
											"receiversUIDs": [
												"63124dffdcf1e4974079a441",
												"63130ba28d796320172e58d6"
											]
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Create a Announcement",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAnnouncements"
							}
						}
					}
				}
			}
		},
		"/announcements/update/{announcementId}": {
			"put": {
				"operationId": "UpdateAnnouncement",
				"responses": {
					"200": {
						"description": "updated",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IAnnouncements"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Update a announcement",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "announcementId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_IAnnouncements_"
							}
						}
					}
				}
			}
		},
...
		"/alarms/{alarmId}": {
			"get": {
				"operationId": "GetAlarm",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IAlarms"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "the requested Alarm in not found"
					}
				},
				"description": "Get a Alarm details",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "alarmId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "63133c886978820e456b8015"
					}
				]
			},
			"delete": {
				"operationId": "DeleteAlarm",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Document_any.any.any_"
										},
										{
											"$ref": "#/components/schemas/IAlarms"
										},
										{
											"$ref": "#/components/schemas/Required_any_"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "the requested alarm in not found"
					}
				},
				"description": "Delete a alarm",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "alarmId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "63133c886978820e456b8015"
					}
				]
			}
		},
		"/alarms/create": {
			"post": {
				"operationId": "CreateAlarm",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAlarms"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "new therapist",
											"referenceType": "Agreement",
											"referenceID": "63131c361ea32f520cb28bd6",
											"frequencyUnit": "Days",
											"frequency": 15,
											"active": true,
											"startDate": "2022-09-10T00:00:00.000Z",
											"endDate": "2022-09-10T00:00:00.000Z"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Create a alarm",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAlarms"
							}
						}
					}
				}
			}
		},
		"/alarms/update/{alarmId}": {
			"put": {
				"operationId": "UpdateAlarm",
				"responses": {
					"200": {
						"description": "updated",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IAlarms"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Update a alarm",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "alarmId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "63133c886978820e456b8015"
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_IAlarms_"
							},
							"example": "63133c886978820e456b8015"
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}